renaming mds-json-schema -> mds-schema-validators and cleaning up build issues
