+++ packages/mds-compliance/mds-compliance-cli.ts
@@ -17,7 +17,7 @@ import * as fs from 'fs'
 import log from '@mds-core/mds-logger'
 import * as yargs from 'yargs'
 import { Policy, Geography, ComplianceResponse, VehicleEvent, Device } from '@mds-core/mds-types'
-import { validateEvents, validateGeographies, validatePolicies } from '@mds-core/mds-json-schema'
+import { validateEvents, validateGeographies, validatePolicies } from '@mds-core/mds-schema-validators'
 import { filterPolicies, processPolicy, filterEvents } from './mds-compliance-engine'
 
 const args = yargs
