+++ packages/mds-compliance/tests/engine-test.spec.ts
@@ -8,12 +8,7 @@ import { la_city_boundary } from '@mds-core/mds-policy/tests/la-city-boundary'
 import { FeatureCollection } from 'geojson'
 import { processPolicy, filterPolicies, filterEvents } from '@mds-core/mds-compliance/mds-compliance-engine'
 import { RuntimeError } from '@mds-core/mds-utils'
-import {
-  ValidationError,
-  validateEvents,
-  validateGeographies,
-  validatePolicies
-} from '@mds-core/mds-json-schema'
+import { ValidationError, validateEvents, validateGeographies, validatePolicies } from '@mds-core/mds-schema-validators'
 
 let policies: Policy[] = []
