+++ packages/mds-schema-validators/tsconfig.build.json
